requirements.txt
APScheduler==3.10.4
feedparser==6.0.11
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3